Song Controller: Extend query params and remove song routes in other controllers
