+++ server/src/search/search.controller.ts
@@ -2,13 +2,7 @@ import {
 	Controller, Get, Param
 } from '@nestjs/common';
 import { ApiOperation, ApiTags } from '@nestjs/swagger';
-import { PaginationParameters } from 'src/pagination/models/pagination-parameters';
-import SongQueryParameters from 'src/song/models/song.query-params';
-import { SongResponseBuilder } from 'src/song/models/song.response';
-import { PaginationQuery } from 'src/pagination/pagination-query.decorator';
-import RelationIncludeQuery from 'src/relation-include/relation-include-query.decorator';
-import SortingQuery from 'src/sort/sort-query.decorator';
-import Response, { ResponseType } from 'src/response/response.decorator';
+import Response from 'src/response/response.decorator';
 import { SearchAllResponseBuilder } from './models/search-all.response';
 import AlbumService from 'src/album/album.service';
 import ArtistService from 'src/artist/artist.service';
@@ -29,7 +23,7 @@ export default class SearchController {
 		summary: 'Search items by their names'
 	})
 	@Response({ handler: SearchAllResponseBuilder })
-	@Get('/all/:query')
+	@Get(':query')
 	async searchItems(
 		@Param('query')
 		query: string,
@@ -40,28 +34,5 @@ export default class SearchController {
 			songs: await this.songService.search(query, {}),
 			genres: await this.genreService.getMany({ slug: { contains: query } })
 		};
-	}
-
-	@ApiOperation({
-		summary: 'Search songs by their names'
-	})
-	@Response({
-		handler: SongResponseBuilder,
-		type: ResponseType.Page
-	})
-	@Get('/songs/:query')
-	async searchSongs(
-		@Param('query')
-		query: string,
-		@PaginationQuery()
-		paginationParameters: PaginationParameters,
-		@RelationIncludeQuery(SongQueryParameters.AvailableAtomicIncludes)
-		include: SongQueryParameters.RelationInclude,
-		@SortingQuery(SongQueryParameters.SortingKeys)
-		sortingParameter: SongQueryParameters.SortingParameter
-	) {
-		return this.songService.search(
-			query, {}, paginationParameters, include, sortingParameter
-		);
 	}
 }
